+++ src/com/universidad/modelo/Persona.java
@@ -1,14 +1,14 @@
 package com.universidad.modelo;
 
 public abstract class Persona {
-    protected Double id;
+    protected Long id;
     protected String nombres;
     protected String apellidos;
     protected String email;
     
     public Persona() {}
     
-    public Persona(Double id, String nombres, String apellidos, String email) {
+    public Persona(Long id, String nombres, String apellidos, String email) {
         this.id = id;
         this.nombres = nombres;
         this.apellidos = apellidos;
@@ -16,8 +16,8 @@ public abstract class Persona {
     }
     
     // Getters y Setters
-    public Double getId() { return id; }
-    public void setId(Double id) { this.id = id; }
+    public long getId() { return id; }
+    public void setId(long id) { this.id = id; }
     public String getNombres() { return nombres; }
     public void setNombres(String nombres) { this.nombres = nombres; }
     public String getApellidos() { return apellidos; }
